예외처리 함수 구현
