+++ README.md
@@ -18,3 +18,4 @@
 3. String으로 받은 입력 값을 Int배열로 바꾸는 함수 구현
 4. 스트라이크와 볼 체크하는 함수 구현
 5. 비교해서 결과를 출력하는 함수 구현
+6. 입력값 예외처리 함수 구현
