+++ build.gradle
@@ -12,8 +12,3 @@ dependencies {
     testCompile('org.junit.jupiter:junit-jupiter:5.5.2')
     testCompile('org.assertj:assertj-core:3.14.0')
 }
-
-compileJava.options.encoding = 'UTF-8'
-tasks.withType(JavaCompile){
-options.encoding='UTF-8'
-}
